feat: move templates logic to another directory
