+++ create-board/routes/routes.go
@@ -1,0 +1,13 @@
+package routes
+
+import (
+	"image"
+
+	tmpl "github.com/milesq/verses-page/create-board/utils/templates"
+)
+
+var templates map[string]image.Image
+
+func init() {
+	templates = tmpl.LoadTemplates()
+}
